Standardise interface of all protocols, and add a queue system to allow partial (de)serialisation
There would have been considerable complexity when using the protocols with the original design, as different protocols had different input and output formats. This standardises all protocols to have the serialised form as a bytes object, and the deserialised form as a tuple of values.

It is envisaged that a chunk of data data received from some hardware interfaces (e.g. serial ports) may not coincide with exactly one, complete packet. This makes it difficult to deserialise received data into values in one go. Instead, now data can be passed in in arbitrary chunks, and the protocol class can manage the handling of said chunks. The first concrete protocol class - StringBinaryProtocol - is partially implemented - it will work with encodings with one-to-one mappings of bytes to characters (like ASCII) but may fail if chunks are split on byte boundaries.